migrate posts to RSC
